# docs/openapi.yaml
openapi: 3.1.0
info:
  title: QuickServe API
  version: 1.0.0
  description: |
    QuickServe API - 地域の熟練ワーカーのためのサービス予約プラットフォーム

    ## 認証
    すべての保護されたエンドポイントは Bearer トークンを必要とします。
    `Authorization: Bearer <token>` ヘッダーを含めてください。
servers:
  - url: http://localhost:3000/api/v1
    description: Development server
  - url: https://api.quickserve.com/api/v1
    description: Production server
...
paths:
  /auth/otp/send:
    post:
      tags: [Auth]
      summary: OTP送信
      description: 電話番号にOTPを送信
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [phone]
              properties:
                phone:
                  type: string
                  example: "0241234567"
      responses:
        '200':
          description: OTP送信成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SuccessResponse'
  /auth/otp/verify:
    post:
      tags: [Auth]
      summary: OTP検証
      description: OTPを検証してトークンを取得
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [phone, code]
              properties:
                phone:
                  type: string
                  example: "0241234567"
                code:
                  type: string
                  example: "123456"
      responses:
        '200':
          description: 認証成功
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  data:
                    $ref: '#/components/schemas/AuthResponse'
...
  /bookings:
    post:
      tags: [Bookings]
      summary: 予約作成
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateBookingInput'
      responses:
        '201':
          description: 予約作成成功
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  data:
                    $ref: '#/components/schemas/Booking'
  /bookings/{id}:
    get:
      tags: [Bookings]
      summary: 予約詳細取得
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: 予約詳細
  /bookings/{id}/status:
    patch:
      tags: [Bookings]
      summary: 予約ステータス更新
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        content:
          application/json:
            schema:
              type: object
              required: [status]
              properties:
                status:
                  type: string
                  enum: [ACCEPTED, WORKER_EN_ROUTE, IN_PROGRESS, COMPLETED, CANCELLED]
      responses:
        '200':
          description: 更新成功
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
    SuccessResponse:
      type: object
      properties:
        success:
          type: boolean
        data:
          type: object
...
    AuthResponse:
      type: object
      properties:
        accessToken:
          type: string
        refreshToken:
          type: string
        user:
          $ref: '#/components/schemas/User'
    User:
      type: object
      properties:
        id:
          type: string
        phone:
          type: string
        email:
          type: string
        firstName:
          type: string
        lastName:
          type: string
        avatarUrl:
          type: string
        role:
          type: string
          enum: [CUSTOMER, WORKER, ADMIN]
        status:
          type: string
          enum: [ACTIVE, INACTIVE, SUSPENDED, PENDING_VERIFICATION]
...
    CreateBookingInput:
      type: object
      required: [categoryId, description, latitude, longitude, address]
      properties:
        categoryId:
          type: string
        workerId:
          type: string
        description:
          type: string
        latitude:
          type: number
        longitude:
          type: number
        address:
          type: string
        scheduledAt:
          type: string
          format: date-time
    Booking:
      type: object
      properties:
        id:
          type: string
        customerId:
          type: string
        workerId:
          type: string
        categoryId:
          type: string
        description:
          type: string
        address:
          type: string
        status:
          type: string
          enum: [PENDING, ACCEPTED, WORKER_EN_ROUTE, IN_PROGRESS, COMPLETED, CANCELLED, DISPUTED]
        estimatedPrice:
          type: number
        finalPrice:
          type: number
        scheduledAt:
          type: string
          format: date-time
        createdAt:
          type: string
          format: date-time
...